Update db queries
