+++ scripts/import-athletes.ts
@@ -2,9 +2,8 @@ import { db } from "../src/db/client";
 import { athletes, athleteResults, athleteHonors } from "@/db/schema";
 import { getAthleteById, worldAthleticsClient } from "@/lib/world-athletics";
 import { countryCodeMap } from "@/lib/utils/country-codes";
-import { like, or, sql, and } from "drizzle-orm";
+import { like, or, sql, and, eq } from "drizzle-orm";
 import { openai } from "../src/lib/openai";
-import { eq } from "drizzle-orm";
 import fs from "fs";
 import path from "path";
 import { slugify } from "@/lib/utils";
