--- src/app/run-clubs/page.tsx
@@ -1,5 +1,0 @@
-import { ComingSoon } from "@/components/coming-soon";
-
-export default function RunClubsPage() {
-	return <ComingSoon />;
-}
